+++ decidim-meetings/app/helpers/decidim/meetings/admin/application_helper.rb
@@ -16,6 +16,14 @@ module Decidim
         def tabs_id_for_service(service)
           "meeting_service_#{service.to_param}"
         end
+
+        def tabs_id_for_agenda_item(agenda_item)
+          "meeting_agenda_item_#{agenda_item.to_param}"
+        end
+
+        def tabs_id_for_agenda_item_child(agenda_item)
+          "meeting_agenda_item_#{agenda_item.to_param_child}"
+        end
       end
     end
   end